integrate modal & api
