+++ src/partials/compounds/review-abtest.hbs
@@ -7,23 +7,47 @@
 		data-url="https://dev.cocoandeve.com/products/sunny-honey-bali-bronzing-self-tan-mousse"
 		data-image-url="//cdn.shopify.com/s/files/1/0286/1327/9779/products/FOAM_DARK_YELLOW_1_large.jpg%3Fv=1609922006"
 		data-description="<div> <strong></strong>Anti-cellulite, anti-ageing self tan with zero nasties. Created using unique CellushapeTM formula to hydrate and firm the skin.<br> </div> <ul> <li>Blurs Pigmentation and Perfects Skin.</li> <li>Tropical Mango and Guava Scent. (No biscuit smells!)</li> <li>Lightweight, non-sticky formula.</li> <li>Fast drying and develops in just 2 hours</li> <li>Vegan. 100% Natural DHA. No Nasties</li> </ul>">
-		<div class="row">
+		<div class="row d-lg-none">
 			<div class="col">
-				<img src="https://cdn.cocoandeve.com/reviews/review-thumb-2115-1.jpg" class="w-100" />
+				<img src="//via.placeholder.com/120x120?text=1" class="w-100" />
 			</div>
 			<div class="col">
-				<img src="https://cdn.cocoandeve.com/reviews/review-thumb-2115-1.jpg" class="w-100" />
+				<img src="//via.placeholder.com/120x120?text=1" class="w-100" />
 			</div>
 			<div class="col">
-				<img src="https://cdn.cocoandeve.com/reviews/review-thumb-2115-1.jpg" class="w-100" />
+				<img src="//via.placeholder.com/120x120?text=1" class="w-100" />
 			</div>
 			<div class="col">
-				<img src="https://cdn.cocoandeve.com/reviews/review-thumb-2115-1.jpg" class="w-100" />
+				<img src="//via.placeholder.com/120x120?text=1" class="w-100" />
 			</div>
 			<div class="col">
-				<img src="https://cdn.cocoandeve.com/reviews/review-thumb-2115-1.jpg" class="w-100" />
+				<img src="//via.placeholder.com/120x120?text=1" class="w-100" />
 				<span class="position-absolute text-primary text-underline">+245</span>
 			</div>
 		</div>
+		<ul class="d-none d-lg-flex list-unstyled">
+			<li>
+				<img src="//via.placeholder.com/120x120?text=1" class="w-100" />
+			</li>
+			<li>
+				<img src="//via.placeholder.com/120x120?text=2" class="w-100" />
+			</li>
+			<li>
+				<img src="//via.placeholder.com/120x120?text=3" class="w-100" />
+			</li>
+			<li>
+				<img src="//via.placeholder.com/120x120?text=4" class="w-100" />
+			</li>
+			<li>
+				<img src="//via.placeholder.com/120x120?text=5" class="w-100" />
+			</li>
+			<li>
+				<img src="//via.placeholder.com/120x120?text=6" class="w-100" />
+			</li>
+			<li class="position-relative mr-0">
+				<img src="//via.placeholder.com/120x120?text=7" class="w-100" />
+				<span class="position-absolute text-primary text-underline">+265</span>
+			</li>
+		</ul>
 	</div>
 </div>
